+++ src/main/java/com/zea7ot/algorithm/summaries/graph/minimum_spanning_tree/KruskalsAlgorithm.java
