+++ src/main/java/com/zea7ot/algorithm/summaries/resources.txt
