+++ src/main/java/com/zea7ot/algorithm/summaries/string/pattern_matching/ZAlgorithm.java
