+++ src/main/java/com/zea7ot/algorithm/summaries/tree/BinaryIndexedTree.java
@@ -9,31 +9,33 @@
  * https://github.com/mission-peace/interview/blob/master/src/com/interview/tree/FenwickTree.java
  * https://en.wikipedia.org/wiki/Fenwick_tree
  */
-package com.zea7ot.algorithm.summary.tree;
+package com.zea7ot.algorithm.summaries.tree;
 
 public class BinaryIndexedTree {
     /**
-     * to construct a Binary Index Tree
+     * to construct a Binary Indexed Tree
      * 
-     * Time Complexity: O(N * lg(N)) for the first time
-     * Space Complexity: O(N)
+     * Time Complexity:     O(N * lg(N)) for the first time
+     * Space Complexity:    O(N)
      */
     public int[] constructBinaryIndexedTree(int[] input){
-        int[] binaryIndexedTree = new int[input.length + 1];
-        for(int i = 1; i <= input.length; i++){
+        final int N = input.length;
+        int[] binaryIndexedTree = new int[N + 1];
+        for(int i = 1; i <= N; i++){
             updateBinaryIndexedTree(binaryIndexedTree, input[i - 1], i);
         }
         return binaryIndexedTree;
     }
 
     /**
-     * to update one element in the Binary Index Tree
+     * to update one element in the Binary Indexed Tree
      * 
-     * Time Complexity: O(lg(N))
-     * Space Complexity: O(1)
+     * Time Complexity:     O(lg(N))
+     * Space Complexity:    O(1)
      */
     public void updateBinaryIndexedTree(int[] binaryIndexedTree, int val, int index){
-        while(index < binaryIndexedTree.length){
+        final int N = binaryIndexedTree.length;
+        while(index < N){
             binaryIndexedTree[index] += val;
             index = getNext(index);
         }
@@ -42,16 +44,16 @@ public class BinaryIndexedTree {
     /**
      * to get the prefix sum
      * 
-     * Time Complexity: O(lg(N))
-     * Space Complexity: O(1)
+     * Time Complexity:     O(lg(N))
+     * Space Complexity:    O(1)
      * 
      * @return the prefix sum
      */
-    public int getSum(int[] binaryIndexeTree, int index){
+    public int getSum(int[] binaryIndexedTree, int index){
         index = index + 1;
         int sum = 0;
         while(index > 0){
-            sum += binaryIndexeTree[index];
+            sum += binaryIndexedTree[index];
             index = getParent(index);
         }
         return sum;
@@ -62,6 +64,9 @@ public class BinaryIndexedTree {
      *  1). to get its complement of 2s                     11111101 + 1 = 11111110
      *  2). to Bitwise AND it with the original number      00000010 & 11111110 = 00000010 (2)
      *  3). to add it to the original number                00000010 + 00000010 = 00000100
+     * 
+     * Time Complexity:     O(1)
+     * Space Complexity:    O(1)
      * 
      * @param index
      * @return the next
@@ -76,8 +81,8 @@ public class BinaryIndexedTree {
      *  2). to Bitwise AND it with the original number      00001010 & 00000110 = 00000010
      *  3). to subtract it from the original number         00001010 - 00000010 = 00001000
      * 
-     * Time Complexity: O(1)
-     * Space Complexity: O(1)
+     * Time Complexity:     O(1)
+     * Space Complexity:    O(1)
      * 
      * @param index
      * @return the parent
